requirements.txt
allure-pytest==2.9.43
allure-python-commons==2.9.43
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
contextlib2==21.6.0
data==0.4
decorator==5.0.9
delayed-assert==0.3.5
Faker==8.12.1
funcsigs==1.0.2
idna==3.2
iniconfig==1.1.1
numpy==1.21.2
packaging==21.0
pandas==1.3.2
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-expect==1.1.0
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
schema==0.7.4
six==1.16.0
text-unidecode==1.3
toml==0.10.2
u-msgpack-python==2.7.1
urllib3==1.26.6
